New Files Added

+++ 8-LeetcodeProblems/ConstructRectangle.java
@@ -9,7 +9,7 @@ class ConstructRectangle
 	}
 	public static int[] constructRectangle(int area) 
 	{
-		int w = Math.sqrt(area);
+		int w = (int)Math.sqrt(area);
 		while(area%w != 0)
 		{
 			w--;
